Attempt to set CMDSUBDOC_F_INSERT_DOC flag in lcb_cmdsubdoc for upsert operations.
This changeset is a WIP/POC attempting to diagnose the issue
with doing a subdoc upsert with fulldoc insert
(see subdoc_t changes). LCB API V4 removes direct access
to the flags.

I pulled in internalstructs.h to facilitate direct
access to the opcode/fulldoc insert/subdoc upsert flags for the
purposes of comparison. One can switch between my attempt
to use the official LCB V4 API and this hack using the
PYCBC_WORKAROUND_SD_BREAKAGE flag (defining it turns on the
hack).

This hack is not intended to be submitted to the codebase.

PYCBC_BYPASS_V3_FAILURES

Change-Id: Idde8760d1c95f7da014b6917c042d1b2b85dc76d