PYCBC-608: Fix segfault due to deallocated key
MOTIVATION
----------
When attaching a tracing scope to a given key
prior to scheduling an operation, we decrement the reference count.

This occasionally leads to dangling references to a key
which has been incorrectly destroyed.

The PYCBC_XDECREF command appears to be a remnant from earlier code
which copied the reference and incremented the reference count.

It doesn't appear to be correct or necessary now.

CHANGES
-------
Remove the command to decrement the key reference count.

RESULTS
-------
We no longer experience dangling references to deleted keys.
Code works as expected.

Change-Id: I5da8e4bae84f8db25ee2a937dd94703b55c31360