PYCBC-663: Rerun intermittently failing touch test
Motivation
----------
The latency between 'touching' a document and this taking effect
can be quite large, causing touch tests to fail. This is also
seen in other client implementations. This causes otherwise
problem-free test runs to fail, potentially hiding other problems.

Until we have a secondary means of confirmation that a touch
command has been received and processed, we should mitigate this
unpredictability.

Changes
-------

Use 'flaky' to rerun the touch tests a number of times and use
a threshold of successful runs.

Results
-------

Tests pass as before, and touch tests seem to be more reliable.

Change-Id: I6b109acf44376f6d23326f769a0f97fad8041978
Reviewed-on: http://review.couchbase.org/115794
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>