PYCBC-708 Several similar admin test failures
Motivation:
==========
Some admin tests failed against 6.0 servers when trying to use groups.
Since that's a 6.5 feature, we should not run those 4 tests against 6.0.
Also, the error handler raised another exception due to how it was processing
the original error - lets fix that too

Modification:
============
Convert the 'value' in the HTTPError to a string if needed - that prevents the
second exception.  Since get_all_groups will 404 if groups are not supported, we
raise a NotSupportedError when we 404 on that endpoint.  Then, we can look for
that in the tests, and skip when appropriate.

Results:
=======
Against a 6.0 server, 4 of the 10 tests in usermgt_t.py now are skipped like they
should be.  Against 6.5, they all run.  Sadly, upsert_group seems to be broken currently
so 2 of the 4 tests fail for _that_ reason still.  We will fix that separately.

Change-Id: Ic9a933572995878ebbcf46a630ea6340bd0088d2
Reviewed-on: http://review.couchbase.org/118980
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ellis Breen <ellis.breen@couchbase.com>