PYCBC-704 - Implement options for CollectionManager
Motivation
==========
To conform to RFC, we need to allow and options object to be used.
Also, it would be nice to allow the kwargs to override some of the
options.  Of course, these options are pretty sparse, just a timeout.
Also, you can specify _just_ the kwargs, and we pass that along.
So you can specify timeout=timedelta(seconds=1.4) and not pass in an
options object at all.

Modification
============
All the options needed the correct base class, and the args needed to use
the existing forward_args we use elsewhere.  Also - we were getting the
time in microseconds wrong -- 10x too many of them.  Simple fix.

Result
======
All existing collection manager tests still pass.  The actual testing of the
kwargs/options forwarding maybe needs to be done more generally than just in
this one place, so didn't mock out the admin_bucket.  Could do that though...

Change-Id: I59fa85e68bc4c2671b75ab281d85106efad12e9d
Reviewed-on: http://review.couchbase.org/119910
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ellis Breen <ellis.breen@couchbase.com>