PYCBC-610 FTS Index Manager
Motivation
=========
A management interface for full-text search indexes.

Modification
===========
Added per the RFC, with associated tests.  Also added a helper to
call a function multiple times, pausing whenever it raises an
exception.  Helpful for testing as somtimes, if you immediately
try to get the document count, or an analysis, we can fail.  Better
to be more deterministic, but this is quick to add.

Results
=======
Tests pass

Change-Id: I4d3280bf7506a4125737ea196e8598534710fd0d
Reviewed-on: http://review.couchbase.org/120332
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ellis Breen <ellis.breen@couchbase.com>