PYCBC-694 Make query interface RFC-compliant
Motivation
==========
Making query match the RFC.

Modification
============
We have the QueryOptions you'd expect from the RFC, and we pass most
of them along.  However, we don't yet actually support the new options,
as we are using the existing N1QLQuery under the hood.  A subsequent
PR will address that.

Result
======
Small set of new tests all pass, and test the basic functionality.  We
will eventually test the entire cluster interface there, but now all
the cluster tests are query tests.  Could be we need to revisit that
decision later, splitting them out, etc...

Change-Id: If704b95e2e78adbae1d3ecb7c94296729b0da886
Reviewed-on: http://review.couchbase.org/120590
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ellis Breen <ellis.breen@couchbase.com>