PYCBC-763: Correct MutateInResult.content_as implementation
Motivation
----------
Unfortunately this method wasn't marked as a property,
which made it usable as a method but makes the syntax ugly
and inconsistent. It should also have returned a
ContentProxySubdoc as per the RFC.

Changes
-------
Made MutateInResult.content_as a property.
Return a ContentProxySubdoc instead of a ContentProxy.
Add a test to verify this works properly.
Fix return type annotation of mutate_in.

Results
-------
All tests passing.

Change-Id: Iec1f8515f6d5e46877e8483ca262a0ac3645096b
Reviewed-on: http://review.couchbase.org/120744
Tested-by: Ellis Breen <ellis.breen@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>