PYCBC-622 Eliminate git dependency in python client build
Motivation
=========
There is no need for a git dependency - locally we can always use
git to get a specific gerrit changeset (in a libcouchbase cloned repo),
and build with that.  I couldn't find an easy way to make pip guarantee git
would exist on a system trying to install the client, so this seems
safer.

Modification
===========
Removed several things.  No more patches.  No more complex decisions
over what tag to use given a specifiy PYCBC_LCB_API.  Lets instead
pin it to the version of the code, and allow you to override with
an environment variable.  Since HYBRID_BUILD is always true, removed
conditionals.

We can bring this stuff back (even the git dependency) within some conditional
statements if we _really_ need it.

Results
======
Builds without Git.  If you want a specific libcouchbase tag, you can
specify it when you build, for instance:
    LCB_TAG=3.0.0 python3 setup.py build_ext --inplace
and a full SHA1 can be used, as well as a tag.  For gerrit patch sets
just build libcouchbase locally with that changeset, and follow the
directions in the README to build the extension.

Change-Id: I512854cd7f60e4c165168a32ef3354887a1a6806
Reviewed-on: http://review.couchbase.org/121698
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Matt Ingenthron <ingenthr@gmail.com>