PYCBC-691 - Get integration tests running on Jenkins
Motivation
==========
Gerrit triggers use the mock, which is great for each commit to
a PR.  However, when we merge to master (or perhaps nightly when
there are changes, or whatever) we should run against real servers.

Modification
===========
Well, mostly that was done on Jenkins, and that work will end up in
the repo that houses the pipelines.  But, we also need to make sure
that the v3 tests _pass_.  This is a first cut at that.  There are
still a few issues, but this is a step forward.

Result
======
Less fails in v3

Change-Id: I284e60f7e4a3522d623ee0d1e19bad67572798c0
Reviewed-on: http://review.couchbase.org/122208
Reviewed-by: Matt Ingenthron <ingenthr@gmail.com>
Tested-by: Build Bot <build@couchbase.com>