PYCBC-691 Integration tests
Motivation
==========
This addresses _some_ of the failures we've seen in jenkins.  More
to come.

Modification
===========
Remove some sleeps, replacing with polling that seems more reliable,
and poll in a number of places we didn't.  Whether or not we got an
error was just a matter of luck, this tightens it up for the collection
management and management tests.

Results
=======
More tests pass, especially on Jenkins against a cbdyncluster.

Change-Id: I5c12a68d06e9f72d1a81c79468840a32a1bc268e
Reviewed-on: http://review.couchbase.org/122284
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ellis Breen <ellis.breen@couchbase.com>