PYCBC-732 QueryIndexManager interface work
Motivation
==========
The interface needs to match the RFC, and the tests should all be passing

Modification
============
It seems that on 6.5 with dev preview on, index creation never really
completes.  Hence one issue with the tests.  So, added a private function
to Cluster that tells you if this is the case, and used that to skip tests that
depend on adding indexes.  Matched interface with RFC, modified tests to
match that.  Also broke tests up so each test tests one thing, so failures
in the future are easier to understand.

Result
======
More passing tests, correct interface

Change-Id: I36e515c75ee173b649936ed4242b19cf565f6066
Reviewed-on: http://review.couchbase.org/122683
Reviewed-by: Ellis Breen <ellis.breen@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>