PYCBC-675 - Result object not what was expected
Motivation
==========
RFC compliance on collection interface

Modification
============
This is just a start, addressing just the get functionality of
collection.  The GetOptions and GetResult were updated, tests added,
eliminated SDK2AsyncResult and similar.  Note there is much more to do
but this shows the direction I'd think we should go.

Also note I discovered a number of issues - see the skipped tests
for the details.

Result
======
Tests pass (with skips), a bit closer to RFC.

Change-Id: I70e0d54eab08536ab66164149d7b52ebc0713d77
Reviewed-on: http://review.couchbase.org/120440
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ellis Breen <ellis.breen@couchbase.com>

 Conflicts:
	couchbase/collection.py