PYCBC-756 - KV has invalid named args
Motivation
==========
The collection interface should match the spec

Modification
============
Eliminated named args in a few places (insert, replace, mutate_in).
Also fixed up some of the naming and a bit of the docstrings as I was
sweeping through.  Found an issue with ping (not returning an array
of EndpointPingReports for each service), and fixed it as I needed it
for the tests. Added tests for more of collection interface.  Still need
more.

Result
=====
More tests in collection_t.py.

Change-Id: Id50715501de0a4fed8bcf93c5342e3783a218380
Reviewed-on: http://review.couchbase.org/122670
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ellis Breen <ellis.breen@couchbase.com>