PYCBC-816, PYCBC-815: Enable txcouchbase core implementation
Motivation
----------
We want to support Twisted with SDK3. Additionally we want
to share lcb_INSTANCEs between collections, as they should
be low-footprint, and not need a new connection for each
collection in a single bucket.

Changes
-------

Various fixes to safety-proof debugging and tracing.
Added Python exception traceback logging in PYCBC_DEBUG mode, allowing
us to see the Python call stack when an exception is (re)thrown from
the C bindings.

Correctly shut down the child tracer when destroying a Tracer.
This is necessary to reset an event that would otherwise
be left uncancelled and prevent the Twisted reactor from
completing on shutdown.

More IOPS debugging code in PYCBC_DEBUG mode.

Add a failsafe target for acouchbase testing
if parameterized behaves unexpectedly.

Results
-------
txcouchbase and acouchbase tests all pass,
apart from text_txconn.BasicConnectionTest.test_bad_event, which we have
skipped for now. This appears to be a Python 3 specific issue as the
same problem doesn't appear on Python 2.7.

Change-Id: I40505081376356e9cf1b27ce5907aab586e467ec