PYCBC-736: gcouchbase support
Motivation
----------
The existing gevent support only works for gevent<1.3,
using methods that are no longer present in later gevent
versions. Unfortunately this means that gcouchbase doesn't
work with Python 3, as gevent 1.3 is the first version
that works with it. We should support later gevents.

Changes
-------
Create a new timer each time callback is called instead of
using unsupported behaviour to do so.

Results
-------
Tests pass

Change-Id: Id6e185a43a5354ec8bdb64c88de37b66949add4c