PYCBC-799: Fix integration test errors
Motivation
==========
All our integration tests should pass.

Modification
============
Some test cleanup, mainly.  Since timeouts happen sometimes during
ping, I protected against that.  Also some other mitigation for timing
issues.  There were durability changes that mean our durable writes only
work on 6.5, so added code to deal with that.  Some things got skipped -
get_all_replicas will intermittently segfault, so skipping it and we
have an issue in JIRA for that.  Minor cleanup of ping (adding the
PingState, and checking for it).  The diagnostics is hanging forever
on < 6.5 clusters - skipping in that case and there's an issue in JIRA
for that as well.  We skip analytics tests in < 6.0 since it was in
dev preview at the time.

We also needed to cleanup the actual jenkins job, as it was not testing
all combinations we want.  Some issues above contributed, also needed
to be much more careful about rebalancing, adding indexes, using the
sample buckets, etc...  This addresses the failures we do see, and the
tests should pass with this code on 5.5.x, 6.0.x, 6.5.0

Results
=======
Green integration tests.  And a few new bug reports.

Change-Id: I2138742439cfed3c3ad3bbffc704951121d9059e