PYCBC-815, PYCBC-816: Make Cluster async-capable
Motivation
----------
We want to make Cluster async-capable, and we can reuse the existing
async wrapping code to do so if it inherits from CoreClient.

Changes
-------
Make Cluster inherit from CoreClient, and self._clusterclient is elided
in favour of 'self'.

Results
-------
All tests pass as before.
We now can run Twisted N1QL and Search queries on a Cluster.

Change-Id: I777e0d0cf904544bd0e84a89746bf7a89f12306b