PYCBC-850: collection kwargs not set by CBCollectionShared
Motivation
----------
When reverting the collections code to using a shared-bucket
approach to collections, the new CBCollectionShared class
accidentally omitted to pass through the name of the
collection to CBCollectionBase, the superclass. This
meant everything was performed on the default collection.

Changes
-------
Pass through the collection name to CBCollectionBase.
Add thorough tests to ensure both scope and collection are
honoured.

Fix up some instances where collection wasn't being
injected by the decorators.

Results
-------
We now operate on the correct collection. All tests passing.

Change-Id: I2115d192de120782cc2407a1700246474aeefb7a