PYCBC-886 Expiry needs to be returned and set properly
Motivation
==========
We want to conform to the RFC

Modification
============
You can specify the desired expiry in any Options that are subclasses
of DurabilityOptionBlock now.  That sets it, and we have tests to
confirm that.  GetOptions has a with_expiry flag, and when set we
retrieve it from the xaddrs ($document.exptime), and set it in the
Result.  There is a with_expiry Spec which is used under the covers
there, and can be used in lookup_in as well.  The bindings got just
the minimum needed for that to work.  Also - if I do a get where all
I have in the options is a with_expiry=True (no project), I'd expect
the get to return the full document.  So the bindings now support
the FULLDOC_GET as well.

Result
======
New set of subdoc tests - minimal but we will add more.  Plus testing
the expiry in collection tests using the GetOptions.

Change-Id: I5e56ecb403b24f2b382f83ed1ec648e96bbf2333