PYCBC-918: Admin still was passing in a bucket name.
Motivation
==========
We can't require a bucket named "default".  Lets just fix this like
we fixed the Cluster connection.

Modification
============
If we _do_ pass a bucket name in when creating the cluster, it goes
into the Admin as well.  But, no pushig it in by default anymore.
This also exposed a bug in LCB, hence we pull in the settings->bucket fix here.

Result
======
No new tests.  We _should_ have one, but it would mean deleting and
recreating the default bucket in our tests.  Manipulating the buckets
already proved difficult to deal with, in terms of timings.  So, for now
I pushed without an explicit test for this.  The issue is, if that bucket
doesn't get created with the same params, we will have other (or all) tests
fail later.  Will address this in the future.

Change-Id: I73063269a77357e04f1dca29c4d071bf23e4f2f1