PYCBC-918: Admin still was passing in a bucket name
Motivation
==========
Should not require a bucket with a specific name exists for our
client to work.

Modification
============
We pass in no bucket now.  However, if you open a bucket, and the
admin connection has not been created yet, we open the admin connection
with that bucket first.  Then use it to open the bucket.

Results
=======
Passing tests against mock and a 6.5 and 6.0 cluster locally.

Change-Id: I73063269a77357e04f1dca29c4d071bf23e4f2f1