PYCBC-918: Admin still was passing in a bucket name
Motivation
==========
Should not require a bucket with a specific name exists for our
client to work.

Modification
============
We pass in no bucket now.  However, if you open a bucket, and the
admin connection has not been created yet, we open the admin connection
with that bucket first.  Then use it to open the bucket.

It isn't clear to me that the bucket _ever_ needs to be there for the admin
functionality to work.  For sure, a pre 6.5 cluster needs an open bucket
for query, search, analytics queries.  But the admin stuff seems to only
call http_request, which doesn't need one.

I'll play around with this more and try to better understand when/if we
need a bucket for the Admin.

Results
=======
Passing tests against mock and a 6.5 and 6.0 cluster locally.

Change-Id: I73063269a77357e04f1dca29c4d071bf23e4f2f1