PYCBC-925 Tests need to ensure the services are up
Motivation
==========
Lets have the tests be sure the expected services are up before
running a test on one of them.  Mainly a stability thing, so we
don't run into flaky integration tests when a test cluster is slow
to come up.

Modification
============
Ping does the job.  However there is a ping bug - it doesn't break
out the Analytics, and instead reports it with the Query.  So, it
isn't perfect.  Lets see how that goes.  Eventually wait_until_ready
will be implemented in LCB, and we can use that instead.

While here, fixed up some intermittently failing integration tests,
and added something to help tests when we move to testing against
7.0 candidates (shortly).

Results
=======
No new tests, hoping they are a bit less likely to have flaky results.

Change-Id: Ibf4ef1e747de4b7c25c645bb6c14d55382487bb4