PYCBC-928: fix bootstrap timeout on Centos
Motivation
----------

Cluster bootstrapping was periodically timing out, but only on
Centos against the Mock. This seems to be due to IPV6 not
being disabled by default, as was the case prior to
http://review.couchbase.org/c/couchbase-python-client/+/127360

Not sure why this issue only shows up on this combination, but
perhaps Centos disables IPV6 by default, and this only causes
issues with a loopback connection. Loopback connections
are significantly different on IPV6 vs IPV4.

Changes
-------

Disable IPV6 by default,
Results
-------

Tests run fine, Centos timeouts disappeared.

Change-Id: I7b8470b5fee0717c4f35570c835d7ff411dc26d2