PYCBC-941: Build against correct OpenSSL version on Windows
Motivation
----------
On Windows, the OpenSSL headers and libraries required to build
libcouchbase with TLS support may not be present on the system.
These are very specific to the Python version, so it would be
most convenient for the build to optionally detect and
download the right ones.

Changes
-------
Move all code/config generation for PYCBC into gen_config.py.
Use the `ssl` module to introspect the OpenSSL version required for
Python. If PYCBC_SSL_FETCH is set to 'ALL' or 'GITHUB', attempt
to pull in the right binary dependencies from the cpython-bin-deps
repository.

Results
-------
When PYCBC_SSL_FETCH is set to GITHUB or ALL, OpenSSL is pulled in
and SSL is built against and passes tests.

Change-Id: Id3fb82f744ec2e32c00cd0719713727eeb91146d